alignment window size dependent on n of sentences
I think the horizontal line in line 14 is not necessary.

The length of the alignment window is now dependent on the number of sentences. If a document is shorter than 20 sentences, all sentences are shown at once. If the document is longer, the scrollable box is shown (with an extended height of 750px).